pyproject.toml
[project]
name = "discord-util-bot"
version = "0.1.0"
description = "Discord dynamic voice channel management bot"
readme = "README.md"
license = "MIT"
requires-python = "==3.12.*"
dependencies = [
    "discord.py~=2.6.0",
    "SQLAlchemy~=2.0.0",
    "alembic~=1.18.0",
    "pydantic-settings~=2.0",
    "asyncpg~=0.30.0",
    "greenlet~=3.0.0",
    "fastapi~=0.115.0",
    "uvicorn[standard]~=0.34.0",
    "python-multipart~=0.0.20",
    "itsdangerous~=2.2.0",
    "bcrypt~=4.2.0",
    "psycopg2-binary~=2.9.0",
    "emoji~=2.14.0",
    "httpx~=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest~=8.0",
    "pytest-cov~=4.0",
    "pytest-asyncio~=0.23",
    "pytest-xdist~=3.5",
    "dpytest~=0.7",
    "mypy~=1.8",
    "ruff~=0.14.0",
    "faker~=37.0",
    "httpx~=0.28.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]

[tool.ruff.lint.per-file-ignores]
"src/web/app.py" = ["B008"]  # FastAPI Depends() is standard pattern
"src/web/routes/*.py" = ["B008"]  # FastAPI Depends() is standard pattern
"src/web/templates/*.py" = ["E501"]  # HTML templates have long lines
"tests/**/conftest.py" = ["ARG001", "E402"]  # pytest fixtures; xdist env setup before imports
"tests/**/test_*.py" = ["ARG002"]  # pytest fixtures used for side effects

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src"]
branch = true
concurrency = ["greenlet", "thread"]
omit = [
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
fail_under = 70
